refactor(web): extract usePlaygroundSession() hook from PlaygroundPage (#453)
PlaygroundPage went from 518 lines down to 386 (−25%). Closes the second hook-extraction follow-up from #660 (the SkillDetailPage equivalent landed in #664).

What the hook owns:

- Queries: useSkill, useSkillPackage, useMyQuota, usePlaygroundChat (Zustand-backed chat state surface).
- Refs: messagesEndRef + messagesScrollRef (scroll anchor + scrollable container), chatInputRef (composer focus + setValue), stickToBottomRef (scroll-stick tracker), closeTimerRef (drawer hover-close debounce).
- Derived: envVarKeys, needsEnvVars, allEnvVarsFilled, envIncomplete, playgroundSnap, isOverLimit, activeDrawer, skillCategory, previewMetadata.
- Local state: envVars dict, pickedModelId, hoverDrawer, pinnedDrawer.
- Effects (4): scroll-listener attach/cleanup, auto-scroll on new messages, Esc-to-unpin-drawer, auto-pin Env when incomplete. Plus the per-skill chat-reset effect (clearChat + envVars + scroll state when skillName changes AND on unmount).
- Handlers (6, all useCallback): handleEnvVarChange, handleSend, handleStarterClick, openHover, scheduleHoverClose, togglePin.

What stays on the page:

- The page owns the JSX layout shell — relative flex container, chat section with the hero+conversation+composer, the rail, the drawer overlay with header + close button.
- The page builds 'starters' (depends only on skillName + t) and 'conversationActive' inline.
- The page builds 'railTabs: PlaygroundRailTab[]' inline since it flips with needsEnvVars — pulling it into the hook would force a re-build in every consumer that didn't care.

Cumulative #453 progress on PlaygroundPage: 813 → 386 lines (−53%) across #660 + this PR. Behavior unchanged; 110 web + 793 backend + 17 sdk tests pass.
